more tests and coverage
