+++ dataBaseMySQL_test.go
@@ -1,11 +1,7 @@
 package main
 
 import (
-	"math/rand"
-	"sync"
-	"sync/atomic"
 	"testing"
-	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 )
@@ -58,53 +54,6 @@ func Test13DBMOneTokenTwice(t *testing.T) {
 	err = tDB.Delete(DEBUGToken)
 	if err != nil {
 		t.Errorf("Can't clear table: %v", err)
-	}
-}
-
-// Concurrent goroutines tries to make new short URL in the same time with the same token (debugging)
-func raceNewToken(db Token, url string, t *testing.T) {
-	DEBUG = true
-	defer func() { DEBUG = false }()
-
-	var wg sync.WaitGroup
-	var succes, fail, cnt, i int64
-	cnt = 4
-	var start sync.RWMutex
-
-	start.Lock()
-	rand.Seed(time.Now().UnixNano())
-
-	racer := func(i int64) {
-		defer wg.Done()
-
-		time.Sleep(time.Duration(rand.Intn(42)) * time.Microsecond * 100)
-		t.Logf("%v Racer %d: Ready!\n", time.Now(), i)
-		start.RLock()
-
-		token, err := db.New(url, 1, newTokenTimeOut)
-
-		if err != nil {
-			t.Logf("%v Racer %d: %v \n", time.Now(), i, err)
-			atomic.AddInt64(&fail, 1)
-			return
-		}
-
-		t.Logf("%v Racer %d: Token for %s: %v\n", time.Now(), i, url, token)
-		atomic.AddInt64(&succes, 1)
-	}
-
-	for i = 0; i < cnt; i++ {
-		wg.Add(1)
-		go racer(i)
-	}
-	t.Logf("%v Ready?\n", time.Now())
-	time.Sleep(time.Second * 2)
-	start.Unlock()
-	t.Logf("%v Go!!!\n", time.Now())
-	wg.Wait()
-
-	if succes != 1 || fail != cnt-1 {
-		t.Errorf("Concurent update error: success=%d, fail=%d, total=%d", succes, fail, cnt)
 	}
 }
